Added most autosplits for bosses and areas
Currently missing are the ones that cannot use AreaID or ProgressionID and they require a sort of 'CutsceneID' if possible to be able to trigger them. These are - Piranhas, Mushroom Rock Road, and Yu Yevon.

Known Issues - Garuda split is broken, it splits in the Dinosaur battle prior to Garuda